Cleaned up and commented the CLR and Main for C++.
Cleaned up and commented the CLR and Main for C++. Added a Common folder at the root to store shared code. Included a String.hpp for handling the convertions between wide chars and normal chars. Known bug with the "-modded" flag and or any arguments passed via Steam. The Bootstrapper will still trigger the Mods.